Medical_Booking_App/requirements.txt
streamlit
streamlit-option-menu
pandas
plotly
google-cloud-bigquery
google-auth
google-auth-oauthlib
python-dotenv
